Wrap to inputgenerator module
