+++ requirements.txt
@@ -1,0 +1,2 @@
+torch==1.4.0
+astor==0.8.1
